+++ src/shared/types/index.ts
@@ -1,2 +1,4 @@
 export * from './entities'
 export * from './config'
+export * from './dashboard'
+export * from './mobile'